requirements.txt
datasets
dialz
numpy
pandas
python-dotenv
torch
transformers